reproify: inline the repro into the test as an @isolated function
Fold the separate .child.ts into the test file as an @isolated repro function
(unicorn-js/isolated-functions enforces it references nothing from module
scope — every dep is imported inside). runReplay materializes
reproduce.toString() to a sibling *.ignoreme.ts (gitignored), runs it in a
child capped at the isolate budget, and deletes it; the single arg is baked in
via JSON.stringify. The .toString() is sanitized back from vitest's SSR
transform (native import + absolute specifiers), failing loudly if a vite
artifact leaks. One self-contained file, lint-guaranteed closed.
